CourtConnectWebsite/Thu-13-16-13-main/backend/requirements.txt
asgiref==3.9.1
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
coverage==7.11.0
cryptography==46.0.3
Django==5.2.6
django-axes==8.0.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
paypalrestsdk==1.13.3
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rest-framework-simplejwt==0.0.2
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
